// src/app/interfaces/mensajes-interface.ts
// Generated by https://quicktype.io

export interface ResultMensajes {
    ok:      boolean;
    msg:     string;
    mensaje: Mensaje[];
}
export interface ResultMensaje {
    ok:      boolean;
    msg:     string;
    mensaje: Mensaje;
}
export interface Mensaje {
    _id:         string;
    titulo:      string;
    descripcion: string;
    fecha:       string;
    usuario:     Usuario;
    estado:      boolean;
}

export interface Usuario {
    _id:    string;
    nombre: string;
}